requirements.txt
psutil
scikit-learn
SQLAlchemy
pandas
click
APScheduler
python-dotenv
google-generativeai